requirements.txt
Flask==2.2.2
flask-pymongo==2.3.0
python-dotenv==1.0.0
werkzeug==2.2.3
pyqrcode==1.2.1
pymongo==3.12.1
pydantic==1.10.12
pypng==0.20220715.0
flask_mail==0.9.1